backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic-settings>=2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic[email]==2.6.1
anthropic==0.43.0
langsmith>=0.1.100
Pillow>=10.2
resend>=2.0
httpx==0.27.0
pytrends==4.9.2
# Social publishing (optional)
tweepy>=4.14
# Video analysis (optional)
opencv-python-headless>=4.9
# Visual similarity search (optional — requires CLIP model)
# open-clip-torch>=2.24
# torch>=2.2
# pgvector>=0.2
# S3/CDN storage (optional)
aioboto3>=12.0
# Structured logging
structlog>=24.0
# Error tracking (optional)
sentry-sdk[fastapi]>=1.40
# Report generation
reportlab>=4.1
# Redis caching
redis[hiredis]>=5.0
# Stripe billing
stripe>=8.0
# Testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.1
httpx>=0.27
aiosqlite>=0.20
factory-boy>=3.3
